Rename Sinaloa modules
